Utterly desouls numbox + moves it to a folder
