+++ code/modules/tgui/tgui_input_number.dm
@@ -14,7 +14,7 @@
  * * min_value - Specifies a minimum value. Often 0.
  * * timeout - The timeout of the numbox, after which the modal will close and qdel itself. Set to zero for no timeout.
  */
-/proc/tgui_numbox(mob/user, message = null, title = "Number Input", default = null, max_value = null, min_value = 0, timeout = 0)
+/proc/tgui_input_number(mob/user, message = null, title = "Number Input", default = null, max_value = null, min_value = 0, timeout = 0)
 	if (!user)
 		user = usr
 	if (!istype(user))
@@ -26,7 +26,7 @@
 	/// Client does NOT have tgui_fancy on: Returns regular input
 	if(!user.client.prefs.read_preference(/datum/preference/toggle/tgui_input))
 		return input(user, message, title, default) as null | num
-	var/datum/tgui_numbox/numbox = new(user, message, title, default, max_value, min_value, timeout)
+	var/datum/tgui_input_number/numbox = new(user, message, title, default, max_value, min_value, timeout)
 	numbox.ui_interact(user)
 	numbox.wait()
 	if (numbox)
@@ -34,7 +34,7 @@
 		qdel(numbox)
 
 /**
- * Creates an asynchronous TGUI text input window with an associated callback.
+ * Creates an asynchronous TGUI number input window with an associated callback.
  *
  * This proc should be used to create numboxes that invoke a callback with the user's entry.
  *
@@ -48,7 +48,7 @@
  * * callback - The callback to be invoked when a choice is made.
  * * timeout - The timeout of the numbox, after which the modal will close and qdel itself. Disabled by default, can be set to seconds otherwise.
  */
-/proc/tgui_numbox_async(mob/user, message = null, title = "Number Input", default = null, max_value = null, min_value = 0, datum/callback/callback, timeout = 0)
+/proc/tgui_input_number_async(mob/user, message = null, title = "Number Input", default = null, max_value = null, min_value = 0, datum/callback/callback, timeout = 0)
 	if (!user)
 		user = usr
 	if (!istype(user))
@@ -57,17 +57,17 @@
 			user = client.mob
 		else
 			return
-	var/datum/tgui_numbox/async/numbox = new(user, message, title, default, max_value, min_value, callback, timeout)
+	var/datum/tgui_input_number/async/numbox = new(user, message, title, default, max_value, min_value, callback, timeout)
 	numbox.ui_interact(user)
 
 /**
- * # tgui_numbox
+ * # tgui_input_number
  *
  * Datum used for instantiating and using a TGUI-controlled numbox that prompts the user with
  * a message and has an input for text entry.
  */
-/datum/tgui_numbox
-	/// Boolean field describing if the tgui_numbox was closed by the user.
+/datum/tgui_input_number
+	/// Boolean field describing if the tgui_input_number was closed by the user.
 	var/closed
 	/// The default (or current) value, shown as a default. Users can press reset with this.
 	var/default
@@ -81,13 +81,13 @@
 	var/min_value
 	/// The time at which the tgui_modal was created, for displaying timeout progress.
 	var/start_time
-	/// The lifespan of the tgui_numbox, after which the window will close and delete itself.
+	/// The lifespan of the tgui_input_number, after which the window will close and delete itself.
 	var/timeout
 	/// The title of the TGUI window
 	var/title
 
 
-/datum/tgui_numbox/New(mob/user, message, title, default, max_value, min_value, timeout)
+/datum/tgui_input_number/New(mob/user, message, title, default, max_value, min_value, timeout)
 	src.default = default
 	src.max_value = max_value
 	src.message = message
@@ -98,32 +98,32 @@
 		start_time = world.time
 		QDEL_IN(src, timeout)
 
-/datum/tgui_numbox/Destroy(force, ...)
+/datum/tgui_input_number/Destroy(force, ...)
 	SStgui.close_uis(src)
 	. = ..()
 
 /**
- * Waits for a user's response to the tgui_numbox's prompt before returning. Returns early if
+ * Waits for a user's response to the tgui_input_number's prompt before returning. Returns early if
  * the window was closed by the user.
  */
-/datum/tgui_numbox/proc/wait()
+/datum/tgui_input_number/proc/wait()
 	while (!entry && !closed && !QDELETED(src))
 		stoplag(1)
 
-/datum/tgui_numbox/ui_interact(mob/user, datum/tgui/ui)
+/datum/tgui_input_number/ui_interact(mob/user, datum/tgui/ui)
 	ui = SStgui.try_update_ui(user, src, ui)
 	if(!ui)
-		ui = new(user, src, "NumboxModal")
+		ui = new(user, src, "NumberInput")
 		ui.open()
 
-/datum/tgui_numbox/ui_close(mob/user)
+/datum/tgui_input_number/ui_close(mob/user)
 	. = ..()
 	closed = TRUE
 
-/datum/tgui_numbox/ui_state(mob/user)
+/datum/tgui_input_number/ui_state(mob/user)
 	return GLOB.always_state
 
-/datum/tgui_numbox/ui_data(mob/user)
+/datum/tgui_input_number/ui_data(mob/user)
 	. = list(
 		"max_value" = max_value,
 		"message" = message,
@@ -134,7 +134,7 @@
 	if(timeout)
 		.["timeout"] = CLAMP01((timeout - (world.time - start_time) - 1 SECONDS) / (timeout - 1 SECONDS))
 
-/datum/tgui_numbox/ui_act(action, list/params)
+/datum/tgui_input_number/ui_act(action, list/params)
 	. = ..()
 	if (.)
 		return
@@ -152,30 +152,30 @@
 			SStgui.close_uis(src)
 			return TRUE
 
-/datum/tgui_numbox/proc/set_entry(entry)
+/datum/tgui_input_number/proc/set_entry(entry)
 		src.entry = entry
 
 /**
- * # async tgui_numbox
+ * # async tgui_input_number
  *
- * An asynchronous version of tgui_numbox to be used with callbacks instead of waiting on user responses.
+ * An asynchronous version of tgui_input_number to be used with callbacks instead of waiting on user responses.
  */
-/datum/tgui_numbox/async
-	/// The callback to be invoked by the tgui_numbox upon having a choice made.
+/datum/tgui_input_number/async
+	/// The callback to be invoked by the tgui_input_number upon having a choice made.
 	var/datum/callback/callback
 
-/datum/tgui_numbox/async/New(mob/user, message, title, default, max_value, min_value, callback, timeout)
+/datum/tgui_input_number/async/New(mob/user, message, title, default, max_value, min_value, callback, timeout)
 	..(user, message, title, default, max_value, min_value, timeout)
 	src.callback = callback
 
-/datum/tgui_numbox/async/Destroy(force, ...)
+/datum/tgui_input_number/async/Destroy(force, ...)
 	QDEL_NULL(callback)
 	. = ..()
 
-/datum/tgui_numbox/async/set_entry(entry)
+/datum/tgui_input_number/async/set_entry(entry)
 	. = ..()
 	if(!isnull(src.entry))
 		callback?.InvokeAsync(src.entry)
 
-/datum/tgui_numbox/async/wait()
+/datum/tgui_input_number/async/wait()
 	return
