+++ code/modules/tgui/tgui_input_text.dm
@@ -14,7 +14,7 @@
  * * multiline -  Bool that determines if the input box is much larger. Good for large messages, laws, etc.
  * * timeout - The timeout of the textbox, after which the modal will close and qdel itself. Set to zero for no timeout.
  */
-/proc/tgui_textbox(mob/user, message = null, title = "Text Input", default = null, max_length = null, multiline = FALSE, timeout = 0)
+/proc/tgui_input_text(mob/user, message = null, title = "Text Input", default = null, max_length = null, multiline = FALSE, timeout = 0)
 	if (!user)
 		user = usr
 	if (!istype(user))
@@ -32,7 +32,7 @@
 				return stripped_input(user, message, title, default, max_length)
 		else
 			return input(user, message, title, default)
-	var/datum/tgui_textbox/textbox = new(user, message, title, default, max_length, multiline, timeout)
+	var/datum/tgui_input_text/textbox = new(user, message, title, default, max_length, multiline, timeout)
 	textbox.ui_interact(user)
 	textbox.wait()
 	if (textbox)
@@ -53,7 +53,7 @@
  * * callback - The callback to be invoked when a choice is made.
  * * timeout - The timeout of the textbox, after which the modal will close and qdel itself. Disabled by default, can be set to seconds otherwise.
  */
-/proc/tgui_textbox_async(mob/user, message = null, title = "Text Input", default = null, max_length = null, multiline = FALSE, datum/callback/callback, timeout = 0)
+/proc/tgui_input_text_async(mob/user, message = null, title = "Text Input", default = null, max_length = null, multiline = FALSE, datum/callback/callback, timeout = 0)
 	if (!user)
 		user = usr
 	if (!istype(user))
@@ -62,17 +62,17 @@
 			user = client.mob
 		else
 			return
-	var/datum/tgui_textbox/async/textbox = new(user, message, title, default, max_length, multiline, callback, timeout)
+	var/datum/tgui_input_text/async/textbox = new(user, message, title, default, max_length, multiline, callback, timeout)
 	textbox.ui_interact(user)
 
 /**
- * # tgui_textbox
+ * # tgui_input_text
  *
  * Datum used for instantiating and using a TGUI-controlled textbox that prompts the user with
  * a message and has an input for text entry.
  */
-/datum/tgui_textbox
-	/// Boolean field describing if the tgui_textbox was closed by the user.
+/datum/tgui_input_text
+	/// Boolean field describing if the tgui_input_text was closed by the user.
 	var/closed
 	/// The default (or current) value, shown as a default.
 	var/default
@@ -86,13 +86,13 @@
 	var/multiline
 	/// The time at which the tgui_modal was created, for displaying timeout progress.
 	var/start_time
-	/// The lifespan of the tgui_textbox, after which the window will close and delete itself.
+	/// The lifespan of the tgui_input_text, after which the window will close and delete itself.
 	var/timeout
 	/// The title of the TGUI window
 	var/title
 
 
-/datum/tgui_textbox/New(mob/user, message, title, default, max_length, multiline, timeout)
+/datum/tgui_input_text/New(mob/user, message, title, default, max_length, multiline, timeout)
 	src.default = default
 	src.max_length = max_length
 	src.message = message
@@ -103,32 +103,32 @@
 		start_time = world.time
 		QDEL_IN(src, timeout)
 
-/datum/tgui_textbox/Destroy(force, ...)
+/datum/tgui_input_text/Destroy(force, ...)
 	SStgui.close_uis(src)
 	. = ..()
 
 /**
- * Waits for a user's response to the tgui_textbox's prompt before returning. Returns early if
+ * Waits for a user's response to the tgui_input_text's prompt before returning. Returns early if
  * the window was closed by the user.
  */
-/datum/tgui_textbox/proc/wait()
+/datum/tgui_input_text/proc/wait()
 	while (!entry && !closed && !QDELETED(src))
 		stoplag(1)
 
-/datum/tgui_textbox/ui_interact(mob/user, datum/tgui/ui)
+/datum/tgui_input_text/ui_interact(mob/user, datum/tgui/ui)
 	ui = SStgui.try_update_ui(user, src, ui)
 	if(!ui)
-		ui = new(user, src, "TextboxModal")
+		ui = new(user, src, "TextInput")
 		ui.open()
 
-/datum/tgui_textbox/ui_close(mob/user)
+/datum/tgui_input_text/ui_close(mob/user)
 	. = ..()
 	closed = TRUE
 
-/datum/tgui_textbox/ui_state(mob/user)
+/datum/tgui_input_text/ui_state(mob/user)
 	return GLOB.always_state
 
-/datum/tgui_textbox/ui_data(mob/user)
+/datum/tgui_input_text/ui_data(mob/user)
 	. = list(
 		"max_length" = max_length,
 		"message" = message,
@@ -139,7 +139,7 @@
 	if(timeout)
 		.["timeout"] = CLAMP01((timeout - (world.time - start_time) - 1 SECONDS) / (timeout - 1 SECONDS))
 
-/datum/tgui_textbox/ui_act(action, list/params)
+/datum/tgui_input_text/ui_act(action, list/params)
 	. = ..()
 	if (.)
 		return
@@ -155,30 +155,30 @@
 			SStgui.close_uis(src)
 			return TRUE
 
-/datum/tgui_textbox/proc/set_entry(entry)
+/datum/tgui_input_text/proc/set_entry(entry)
 		src.entry = entry
 
 /**
- * # async tgui_textbox
+ * # async tgui_input_text
  *
- * An asynchronous version of tgui_textbox to be used with callbacks instead of waiting on user responses.
+ * An asynchronous version of tgui_input_text to be used with callbacks instead of waiting on user responses.
  */
-/datum/tgui_textbox/async
-	/// The callback to be invoked by the tgui_textbox upon having a choice made.
+/datum/tgui_input_text/async
+	/// The callback to be invoked by the tgui_input_text upon having a choice made.
 	var/datum/callback/callback
 
-/datum/tgui_textbox/async/New(mob/user, message, title, default, max_length, multiline, callback, timeout)
+/datum/tgui_input_text/async/New(mob/user, message, title, default, max_length, multiline, callback, timeout)
 	..(user, message, title, default, max_length, multiline, timeout)
 	src.callback = callback
 
-/datum/tgui_textbox/async/Destroy(force, ...)
+/datum/tgui_input_text/async/Destroy(force, ...)
 	QDEL_NULL(callback)
 	. = ..()
 
-/datum/tgui_textbox/async/set_entry(entry)
+/datum/tgui_input_text/async/set_entry(entry)
 	. = ..()
 	if(!isnull(src.entry))
 		callback?.InvokeAsync(src.entry)
 
-/datum/tgui_textbox/async/wait()
+/datum/tgui_input_text/async/wait()
 	return
