+++ tgstation.dme
@@ -3833,8 +3833,8 @@
 #include "code\modules\tgui\tgui.dm"
 #include "code\modules\tgui\tgui_alert.dm"
 #include "code\modules\tgui\tgui_input_list.dm"
-#include "code\modules\tgui\tgui_numbox.dm"
-#include "code\modules\tgui\tgui_textbox.dm"
+#include "code\modules\tgui\tgui_input_number.dm"
+#include "code\modules\tgui\tgui_input_text.dm"
 #include "code\modules\tgui\tgui_window.dm"
 #include "code\modules\tgui\states\admin.dm"
 #include "code\modules\tgui\states\always.dm"
