+++ tgui/packages/tgui/interfaces/Input/ListInput.js
@@ -5,12 +5,12 @@
  */
 
 import { clamp01 } from 'common/math';
-import { useBackend } from '../backend';
-import { Box, Button, Section, Stack, Input } from '../components';
+import { useBackend } from '../../backend';
+import { Box, Button, Section, Stack, Input } from '../../components';
 import { KEY_DOWN, KEY_UP, KEY_ENTER } from 'common/keycodes';
-import { Window } from '../layouts';
+import { Window } from '../../layouts';
 import { Component, createRef } from 'inferno';
-import { globalEvents } from "../events";
+import { globalEvents } from "../../events";
 
 export class ListInput extends Component {
   constructor() {
